Modify: Moved BaseTimeEntity between packages
