+++ matgpt/src/main/java/com/ktc/matgpt/chatgpt/entity/GptReview.java
@@ -1,7 +1,6 @@
 package com.ktc.matgpt.chatgpt.entity;
 
-import com.ktc.matgpt.feature_review.review.entity.BaseTimeEntity;
-import com.ktc.matgpt.store.Store;
+import com.ktc.matgpt.utils.BaseTimeEntity;
 import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
